Split out Geometry test cases to own files
This makes each test class inhabit its own file. We import the "base"
geometry test class in such a way that the unittest does not think we
want to test the base class multiple times.
